Reorganize repository and rewrite install script
Previously, the repository was organized by the types of resources.
Files in lib/ were to be symlinked; files in source/ were to be sourced
by other bash files; things with no other home were in misc/.

This organization meant that related things were forced to live apart in
the repository, while the organization itself offered little benefit.

Now, the repository is organized by the different tools.
Bash files are in bash/; ZShell files in zsh/; Vim files vim/; things
with no home are still in misc/.

This reorganization meant the installer had to be rewritten at it can no
longer auto-discover resources to be installed.
Now, it is directly aware of each of the tools that it is installing.
A nice consequence of this is that it is able to provide a more robust
installation experience, complete with backup and viewing capabilities
of existing resources.