Add flag to show search info in java-deadcode
If --verbose is set, the ripgrep search command and results will be
included in the output.

It may be better to print this information to stderr so it can be
redirected to a file, but that may lead to mismatched output where the
stdout and stderr streams are not necessarily kept in sync.