Install ivy
Ivy replaces the builtin completion behavior.

This is completion that happens at the bottom of the screen in a
little popup buffer, not text completion a la "tex<TAB>". For
instance, when you do find-file or switch-buffer and it prompts you at
the bottom of the screen.

Ivy supports integration with recentf to show recent files in the
switch-buffer window, so I no longer need a custom keybind to bring up
the recentf window!