Install avy for cursor jumping
Use avy to jump your cursor to any visible character in the
frame.

This is useful when you have multiple windows open and you want to
edit some text in the other window, and you can see the text you want
to edit.

In that case, hit C-: and then the character you want to jump to, and
a subsequent "tree" prompt will appear, allowing you to navigate to
the matching character.

See the Ivy info node for more help.