[emacs] Hide emphasis markers in org mode
And reorder the file so packages are at the bottom of the file. This
order makes it easier to modify the file, since the most common
operation is to add another package, you just need to add to the end
of the file rather than insert in the middle of existing configuration.