[emacs] Load projectile immediately
The use of :after caused projectile to not be loaded/activated upon
starting Emacs. Presumably this was because ivy and magit are not
necessarily loaded/activated yet.

The addition of :after was simply meant to explicitly declare that the
projectile configuration used variables from those other packages; it
was not intended to alter behavior.