[emacs] Condition on IS-WINDOWS instead
Flip the perspective on what is the default/primary.

Emacs is primarily for work development, which means Mac.
The fact that I'm trying to use it on Windows is second to that.

As such we want to assume Mac, and then condition things on if we're
running on Windows. This may not change the number of conditionals
necessary for packages and configuration, but it better reveals the
intention.