[gitignore] Ignore SOURCEME files
These `SOURCEME` files are a little pattern I invented to be placed in a
directory to enable CLI functionality for that project.

For instance, if I have a CLI project cloned locally and I want to use
its bin/ executables or tab completion without permanently adding it to
my PATH/bashrc/etc. I can create a `SOURCEME` file that enables this
functionality to be sourced locally in a shell session via `. SOURCEME`.