Pre-load git commit message with ticket number
But if there are any flags, then just behave like normal. This is useful
when doing `gc --amend` for example.

Use "!" in the git editor command to make vim start the cursor at the
end of the line/buffer so you don't have to manually move it past the
pre-loaded ticket number.