Script to check for unused Homebrew packages
The only packages that should show up are ones that you invoke yourself,
such as git, mvn, htop, etc.

If there's a package that you don't recognize then it was installed as a
dependency of some other package which has since been removed, so it's a
dead transient that can be removed.

The only exception to this is the bash-completion package which is used
by Bash/.bashrc.