[git] Browse repo at current branch
Include the current branch name in the repo URL so you don't have to
navigate to the matching branch in case that's what you're after, which
seems like it is most of the time.

Plus it's easier to navigate to the default branch when you're on Github
than it is to start on the default and go the other direction.